deepResearchAndPptCreator/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
gradio>=4.12.0
python-dotenv>=1.0.0
groq>=0.4.2
google-search-results>=2.4.2
python-pptx>=0.6.21
pydantic>=2.4.2
typing-extensions>=4.8.0
starlette>=0.27.0
anyio>=3.7.1
click>=8.1.7
httpx>=0.25.0
python-multipart>=0.0.6 